WAF build: delete softlinks only if they point to different location

+++ linux/ws_main.py
@@ -619,7 +619,7 @@ def build_type(bld,build_obj):
 
 
 def post_build(bld):
-    print("copy objects")
+    print("updating softlinks")
 
     exec_p ="../scripts/"
 
@@ -638,16 +638,22 @@ def build(bld):
 
 def build_info(bld):
     pass;
-      
+
+def compare_link(link_path, dst_path):
+    return os.path.abspath(os.readlink(link_path)) == os.path.abspath(dst_path)
+
 def install_single_system (bld, exec_p, build_obj):
     o=bld.out_dir+'/linux/';
     src_file =  os.path.realpath(o+build_obj.get_target())
     if os.path.exists(src_file):
         dest_file = exec_p +build_obj.get_target()
+        relative_path = os.path.relpath(src_file, exec_p)
         if os.path.islink(dest_file):
+            if compare_link(dest_file, relative_path):
+                return
             os.unlink(dest_file)
         if not os.path.lexists(dest_file):
-            relative_path = os.path.relpath(src_file, exec_p)
+            print('{0} --> {1}'.format(build_obj.get_target(), relative_path))
             os.symlink(relative_path, dest_file);
